building: EXE: stop appending .notanexecutable suffix
Stop appending the .notanexecutable suffix to temporary/intermediate
executable file. The processing steps that need to modify the
executable are now all wrapped in the `_retry_operation` helper,
which should handle the interference from anti-virus programs. And
if that is not the case, we should be getting feedback about it
sooner rather than later.