Fix pkgutil.iter_modules for symbolically linked packages
When a package is symbolically linked out of the _MEIPASS directory
(for example, in macOS .app bundles when package contains data files),
there is a path mismatch causing it to get skipped.

This resolves the real path of the package paths and the module paths
so that they can be compared correctly. It changes the comparison to
run in path space rather than module dotted space. I believe this is
equivalent because all modules in the PyiFrozenImporter should be in
the _MEIPASS directory.