tests: macOS: add a test for pkgutil.iter_modules and package with data file
Add a test for `pkgutil.iter_modules` and package with data file
in macOS .app bundles. In such cases, the package's directory is
fully collected into `Contents/Resources` directory, and a symbolic
link to the package's directory is created in `Contents/Frameworks`
(the `_MEIPASS` directory). When validating the package path prefix,
our `pkgutil.iter_modules` implementation needs to account that
resolved package path does not directly match the `_MEIPASS` path
due to the symbolic link.