bootstrap: add .egg directories in sys._MEIPASS to sys.path
It seems that `pkg_resources` is able to discover metadata
collected from eggs as long as the parent directory of the
.egg directory is in `sys.path`. This is automatically true in
our case, because we collect the  `package_name-version.egg`
directories (that contain `EGG-INFO` metadata directories) into
`sys._MEIPASS`.

In contrast, `importlib.metadata` does not seem to pick up
such metadata unless the `.egg` directories themselves are
in `sys.path`. So have the bootstrap script scan for directories
with `.egg` suffix and append them so `sys.path`.