loader: compatibility fix for importlib.resources and namespace packages
In the case of a PEP-420 namespace package, the attribute
`submodule_search_locations` in the spec returned by the
`PyiFrozenImporter.find_spec` must be an instance of the
`importlib._bootstrap_external._NamespacePath` instead of a plain
locations list. This ensures that the path is properly processed
by the code involved in trying to look up the namespace package's
resources via `importlib.resources.files()`.