loader: fix the `path_finder` arg of the `_NamespacePath`
Fix the `path_finder` argument passed to the constructor of
`importlib._bootstrap_external._NamespacePath` that we
construct for the namespace packages. The argument should not
be an object instance, but rather a bound method that retrieves
the spec (so we need to bind our `find_spec` implementation via
a lambda).

The incorrect type of `path_finder` causes an error during
the refresh codepath that is triggered by cache invalidation via
`importlib.invalidate_caches()`; therefore, extend the
`test_importlib_resources_namespace_package_data_files` with
cache invalidation to have this scenario covered by test suite.