compat: resolve importlib.metadata vs importlib_metadata
Import `importlib.metadata` as `importlib_metadata` if stdlib
version is recent enough (python >= 3.10). Otherwise import
the `importlib_metadata`, but explicitly verify the version
of `importlib-metadata` dist, which must be >= 4.6. This
version is needed for extended functionality, such as `group`
and `name` keyword arguments having been added to the
`entry_points()` function.

Add `importlib-metadata >= 4.6` to requirements if python < 3.10.

Add `packaging` >= 20.0 to requirements; used for parsing of
version and requirements strings.