tests: linux: disable sandboxing in QtWebEngine tests if necessary
On linux systems with glibc >= 2.34, QtWebEngine helper process
crashes with SIGSEGV due to use of `clone3` syscall, which is
incompatible with chromium sandbox (see QTBUG-96214). The issue
was fixed in Qt5 5.15.7, however even the latest PyPI wheels
of PySide2 (5.15.2.1) and PyQt5/PyQtWebEngine (5.15.6) still
seem to ship Qt5 5.15.2 (which was probably last publicly
available linux build from the Qt itself).

If we encounter incompatible combination of glibc and Qt5 (for
example, using PyPI wheels under Ubuntu 22.04), we disable the
sandbox, which allows us to perform basic functionality test.

Note that sandboxing is disabled only in corresponding tests;
this has no impact on user-made applications that use QtWebEngine,
where it is up to user to disable the sandboxing if necessary
(i.e., even to get the original, unfrozen, version of their
program running).