hookutils: implement replacement for is_module_satisfies
Replace `is_module_satisfies` with `check_requirement`, which
uses `packaging` to parse the requirement and check the version
against the optional specifier(s), and `importlib.metadata` to
query the dist version.

Keep `is_module_satisfies` as an alias for `check_requirements`
to maintain compatibility with existing hooks (from the contrib
hooks repository) that call the function with only the
`requirement` string. The calls with additional `version` and
`version_attr` arguments will not work anymore, nor will the
fall back to checking module's version when the specified
requirement is not satisfied.

Have the @requires test decorator use `check_requirement`.

Update hooks in this repository to use `check_requirement`
instead of `is_module_satisfies`.

Update Qt hook utilities, where we need to account for the
fact that `check_requirement` does not attempt to fall back
to checking module's version (which in this case was
undesirable and had to be worked around).