building: fix executable-bit preservation for collected shell scripts
Add special handling for data files with executable bit set (such
as shell scripts).

In `COLLECT` and `BUNDLE`, we now set mode `0755` if typecode
matches the one for binary type OR if the source file has an
executable bit set.

In `PKG`, we manually force the PKG typecode for DATA entries to
`b` ("binary") if source file has an executable bit set.