analysis: ensure __main__ is always among excludes
Ensure that `__main__` is always excluded from the modulegraph,
to prevent accidentally pulling `PyInstaller` itself into the modulegraph.
This seems to happen on Windows, because modulegraph is able to resolve
`__main__` as `.../PyInstaller.exe/__main__.py` and analyze it.

The `__main__` has a different meaning during analysis compared
to the program run-time, when it refers to the program's entry-point
(which would always be part of the modulegraph anyway, by virtue
of being the starting point of the analysis).