building: do not raise errors on Info.plist missing from .framework bundles
Lower the severity of a missing `Info.plist` file in a collected
macOS .framework bundle from an error to a warning (unless strict
collection mode is enabled). Turns out that PyQt shipped PyPI
wheels that contain Qt .framework bundles without `Info.plist`
files up until v5.14.1. So be more tolerant to cases like that,
especially since the user might be interested in a regular POSIX
application instead of an .app bundle, or they might not plan to
code-sign the generated .app bundle.