bootloader: cleanup multipackage dependency extraction code
Clean up and document the multipackage dependency extration code
in an attempt to make it easier to read.

When determining the parent directory of the current exectuable
(the former `homepath_parent`, now `this_executable_dir`), use
`executablename` member of `ARCHIVE_STATUS` instead of `homepath`,
as the latter is affected by the contents-directory option while
the former is not. (And this seems to work for onefile builds
because `homepath` there is also affected by contents-directory,
even through it should be not).

Remove the useless else-if branch, whose condition is now
identical to that of the preceding if condition. It was very likely
unused even in its original form, where it attempted to find the
dependency by jumping to the parent directory of the directory
containing the executable, and then applying relative path to the
other executable. However, this relative path should already
encode the relative relation between the executables (accounting
for all four possible combinations regaring build types), and as
such jumping one directory higher would never work.

Fix the path check for the external .PKG archive (side-load mode).
Such archive is located next to the executable, and is not a subject
to contents-directory setting.