bootloader: apply contents-directory to homepath only in onedir mode
Append contents-directory path to homepath member of ARCHIVE_STATUS
only in onedir mode (i.e., if we do not need to extract the
archive).

This implies that we need to check whether extraction is required
when we open the archive (before homepath is set), so move the
corresponding code into `pyi_arch_setup`.