building: fix randomization of binary dep. analysis search paths
De-duplicate the search paths via `list(dict.fromkeys(...).keys())`
instead of `list(set(...))` to ensure that the order of search
paths is preserved, instead of randomized on each run.

This fixes issues stemming from `sys.base_prefix` not being searched
first on Windows; either for python shared library, or for
`VCRUNTIME140.dll` or `VCRUNTIME140_1.dll`. For example, in an
application using both `pywin32` and `PyQt5`, the `pywintypes3X.dll`
references `VCRUNTIME140_1.dll`. If we search `sys.base_prefix`
first (as we should), we will pick up a copy from there, and collect
it into top-level application directory, where `pywintypes3X.dll`
will automatically find it during run-time. But if we search
the `PyQt5/Qt5/bin` directory first (one of additional search paths),
we will pick up the Qt's copy, and collect it while preserving its
parent directory; the `pywintyptes3X.dll` will then find it only
if `PyQt5` python modules are loaded before `pywin32` ones, which
is not necessarily the case.

Note that in the first case, we pick up the Qt-provided copy of the
`VCRUNTIME140_1.dll` in addition to python-provided one, because the
collected Qt shared libraries that refer to it are located in the
same directory,  and referring binary's location takes precedence
over additional search paths.