bootloader: macOS: do not unset DYLD_* environment variables
On macOS, the bootloader is currently explicitly un-setting the
DYLD_* environment variables, but only in onefile mode. As stated
in the comment above the offending code block, we do not rely on
DYLD_* environment variables, but instead modify the library
search paths in collected binaries. By extension, we should be
able to not care about DYLD_* environment variables, and should
not touch them at all.