bootloader: cleanup: replace pyi_utils_set_environment
The `pyi_utils_set_environment()` is now no-op on both Windows
and macOS, and on other OSes it is essentially a one-line call to
`set_dynamic_library_path`. So replace `pyi_utils_set_environment`
with direct calls to `set_dynamic_library_path` (and rename the
latter to `pyi_utils_set_library_search_path`) to make it more
obvious what is going on (and on which OSes).