docs: fix documentation for python interpreter run-time options [skip ci]
Update the argparse description for the `--python-option` command
line option to list all options and refer user to the corresponding
section in the manual.

Fix the examples in the said "Specifying Python Interpreter Options"
section in the manual. The X-options in the example should be
`utf8` and `dev` instead of `utf8_mode` and `dev_mode`, to match
the corresponding python's X-options.