hooks: update PySide6.QtHttpServer hook for PySide6 6.5.3
Explicitly add `PySide6.QtConcurrent` to hidden imports of
`QtHttpServer` to fix missing module error in basic import test
with PySide6 6.5.3 on Windows.

On other OSes, the `QtConcurrent` dependency seems to be
automatically  inferred due to the  `QtHttpServer` extension being
linked against the  `Qt6Concurrent` shared library. But the explicit
hidden import does not hurt in that case, either.