bincache: update cache index after macOS-specific binary processing
Due to an oversight, the cache index is currently updated before
the macOS-specific binary processing. So if this part raises an
error (either path rewriting fails, or binary architecture
validation raises an exception), the subsequent build attempts that
do not clear the cache (via the `--clean` option) end up picking the
partially-processed binary from the cache.

So move the update of cache index to the very end of the
`process_collected_binary` function, where it should have been
in the first place.