rthooks: pkg_resources: normalize path using os.path.normpath
Use `os.path.normpath` instead of `os.path.abspath` to clarify
the intent (the path should already be absolute, but it might
contain '..' elements that we want to get rid of).

Also add a comment as to why we can avoid resolving symbolic links
and why we do not have to explicitly handle alternate prefix
in macOS .app bundles (Contents/Frameworks vs Contents/Resources).