hookutils: qt: extend the search for OpenSSL DLLs on Windows
Extend the search for OpenSSL DLLs for QtNetwork to cover both
32-bit and 64-bit variant (depending on the environment), and
to cover OpenSSL 3.0.x that is used with contemporary Qt6 (and
in some cases, even Qt5).

We now use `bindepend.resolve_library_path` to search for the
DLLs, so we search the default Qt shared libraries directory and
fall to additional search paths (e.g., PATH).