hookutils: qt: implement dependency validation for Qt plugins
Scan the dependency of Qt plugins and automatically exclude plugins
where:
- a Qt shared library dependency is missing
- a Qt shared library dependency is resolved outside of the
  Qt shared library directory

The first condition prevents us from collecting plugins that are
defunct in the build environment.

The second condition prevents plugins from missing Qt dependencies
from pulling those dependencies from unrelated Qt environments
that (in)conveniently happen to be in the search path.

For example, PyQt5 PyPI wheels ship several plugins with missing
dependencies (some of which are in optional add on packages),
and those might end up pulling in system-installed Qt libraries
(linux), Homebrew installation (macOS), or an arbitrary Qt
build that happens to be in PATH (Windows).