hookutils: qt: search and collect OpenSSL dylibs on macOS
Extend `collect_qtnetwork_files` to collect OpenSSL dynamic
libraries on macOS. After extending the QtNetwork test to check
whether `openssl` TLS backend is available, we can see failures
with python 3.8, 3.9, and 3.10 when running with latest PyQt6
and PySide6, which both use Qt 6.6.0. Those python versions likely
provide OpenSSL 1.1.x, while Qt was built with 3.0.x.

On the CI runner, OpenSSL 3.0.x is likely available from Homebrew,
so the tests sees the `openssl` TLS  backend available in the
unfrozen python + Qt, and thus is not skipped. When running the
frozen application, the POSIX build passes, but .app bundle does
not; because the former can access the Homebrew copy of OpenSSL
at run-time, while the latter cannot.

Which means that we should be attempting to collect the OpenSSL
dylibs for QtNetwork on macOS as well.