rthooks: PyQt6/PySide6: add sys._MEIPASS to DYLD_LIBRARY_PATH
Have PyQt6 and PySide6 run-time hooks add sys._MEIPASS to
DYLD_LIBRARY_PATH when not running as a macOS app bundle. This
is required for QtNetwork in Qt6 to be able to pick up OpenSSL
dynamic libraries for its `openssl` TLS backend; this both ensures
that the backend is functional if OpenSSL libs were collected, and
also prevents fallback to external locations, such as Homebrew.

Applicable only to Qt6 and its bindings; macOS Qt5 builds seem to
use the OS-provided SecureTransport API instead of OpenSSL.