modulegraph: remove modulegraph.zipio module
Aside from its corresponding test suite, the only use of
functionality from this module was `zipio.listdir` in
`_find_all_submodules`, which seems to be called by
`_import_importable_package_submodules` to deal with
`from ... import *` statements.

As we do not support zipped eggs anymore, this should be
superfluous. If it turns out to break "portable" python on
Windows (which has stdlib modules in a .zip file), we can
bring back just the minimal required part of the code.