tests: strict pinning of packages with Qt shared library
In addition to `PyQt6` bindings packages, pin the `-Qt6` packages
that provide Qt shared libraries (e.g., `PyQt6-Qt6`,
`PyQt6-QtWebEngine-Qt6`, etc.). This helps us avoid issues when
the newer version of Qt6 libraries has incompatible ABI with the
pinned bindings version.

For the sake of consistency, pin the `PySide6-Addons` and
`PySide6-Essentials` in addition to `PySide6`.

We do not pin the `-Qt5` packages corresponding to `PyQt5` bindings
packages. This is because Qt5 is not actively developed anymore, and
thus Qt 5.15.x has a stable ABI. Plus, it seems that `*-Qt5` 5.15.2
wheels are available for all platforms, while 5.15.11 are available
only for macOS (and some of them only for arm64). So instead of
complicating the `requirements-libraries.txt`, just leave them
unpinned.