pre-safe-import-hooks: gracefully handle cases when six is unavailable
In pre-safe-import hooks that deal with `six` (or vendored `six`),
gracefully handle the cases when the `six` module is unavailable.

In general, the pre-safe-import hooks might end up being called
for non-existent modules, and the hooks need to account for that
(even though the behavior seems to depend on how the missing
module is imported).

For example, a program with a duplicated import,

```
import six.moves
import six.moves
```

seems to end up triggerring the `six.moves` pre-safe-import
hook on the second import. The same happens in more complex
scenarios where these imports are scattered across different
modules.