bindepend: optimize binary-vs-data classification on linux
Avoid running `objdump` on every file, as that becomes prohibitively
costly for projects with a lot of (data) files.

First, perform a check for 4-byte ELF header to filter out most
(if not all) data files; only then, validate the likely binary
using `objdump` to ensure that it is a valid ELF file.