bindepend: suppress ldd warnings about lack of executable permissions
On linux, shared libraries typically have executable permissions
set, and `ldd` displays a warning when this is not the case.
Naturally, shared libraries shipped in PyPI wheels do not have
executable bits set, which results in slew of `ldd` warnings during
our binary dependency analysis. As the lack of executable permissions
does not affect the analysis, suppress the warnings in the output.