tests: split off the base part of requirements-tools.txt
Split the base part of `requirements-tools.txt` into a separate
`requirements-base.txt`; this includes the requirements
that we expect to be also available for pre-release python
versions.

This allows us to set up basic CI for pre-release python early
in that python version's development cycle, when some of the
additional requirements for basic tests might not be yet available
in the form of binary PyPI wheels (e.g., `pywin32`, `lxml`, and
`xattr` with its `cffi` dependency).