bindepend: fix error when calling get_imports() with pathlib.Path
Fix an error when calling `PyInstaller.depend.bindepend.get_imports`
with a `filename` that is a `pathlib.Path` (or rather, not a `str`)
on Windows. This particular error is caused by trying to call
`lower()` under assumption that `filename` is an instance of `str`;
so do the explicit cast to `str` for that part.