docs: add a "Common Issues and Pitfalls" chapter to docs [skip ci]
Add a new documentation chapter, "Common Issues and Pitfalls" which
at the moment covers the following topics:

Frozen applications on macOS and Linux making use of symbolic
links since PyInstaller 6.0, and the implications for distribution
(e.g., taking care that symlinks are preserved when copying the
frozen application or creating an archive).

Launching external programs from the frozen application; provides
the explanation of why launching external applications requires
manual sanitization oflibrary search paths, and how to achieve it.
Primary goal is to document the modifications of libary-search-path
related environment variables - both by bootloader, and potentially
also by run-time hooks. But we also need to mention that on
Windows, `SetDllDirectory()` is used, and that in order to undo
its effects, `SetDllDirectory(NULL)` call is needed...

Multi-processing; document and explain the need for calling
`multiprocessing.freeze_support` when using `multiprocessing`
functionality.

Windowed/noconsole mode on Windows; explain that in noconsole
builds on Windows, standard I/O file handles are `None`, and
provide a work-around for code that naively tries to access
their attributes without first checking that they are available.