building: refactor process_collected_binary
Refactor `process_collected_binary` helper, with goal of removing
the recursive call to itself when both strip and upx are enabled.

This makes the processing flow easier to follow, and makes it
easier to add additional steps (for example, code-signing on
Windows), without having to worry about those steps being
performed in intermediate recursive calls.