rhtooks: qt: create embedded qt.conf if it does not exist
Have the run-time hooks for `PySide2`, `PySide6`, `PyQt5`, and
`PyQt6` check for the presence of embedded `:/qt/etc/qt.conf`
resource, and if not present, inject their own version. This
aims to ensure that the bundled Qt is always relocatable.

All Qt bindings packages, when packaged as PyPI wheels, inject
embedded `qt.conf` file as Qt embedded resource during their
initialization, and use it to set the `PrefixPath` to the
top-level Qt directory within the package's directory. This
enures that all paths in `QLibraryInfo` are properly set, and
is essential for making the package and its bundled Qt
relocatable.

Qt bindings packaged by other providers (for example, Anaconda
packages on all OSes, or packages provided by linux
distributions) do not seem to use this mechanism when running
under unfrozen python.

However, when collected into PyInstaller-frozen application,
`PyQt5` and `PyQt6` packages, even if originally not packaged
as PyPI wheels, seem to enable the embedded `qt.conf` injection,
making themselves relocatable.

On the other hand, `PySide2` package does not, which makes it
non-relocatable. For example, when using Fedora-packaged
`PySide2` in a frozen application, the `PrefixPath` remains
set to `/usr`, and thus a `QtWebEngine`-enabled frozen application
ends up using system-installed `QtWebEngineProcess` helper in
`/usr/lib64/qt5/libexec`, if available (and failing to start if it
is unavailable). In the case of Anaconda-packaged `PySide2`, the
`PrefixPath` path ends up being `/`, and the collected
`QtWebEngineProcess` fails to be discovered.

Therefore, if the embedded `qt.conf` resource is not provided by
the package itself, we attempt to provide our own, to ensure that
the package is relocatable, and that paths in `QLibraryInfo` are
properly set (i.e., that they match the relocatable PyPI wheel
layout, which we are trying to reconstruct in our Qt collection
approach).