hookutils: qt: fix collection of QtWebEngineProcess helper ...
... from Anaconda PySide2 package on Windows

When collecting `QtWebEngineProcess` helper from `PySide2`/`PySide6`
on Windows, always force the target collection directory to be the
PySide package directory (i.e., the relative target collection
directory should be `.`).

This ensures that the helper is collected in PyPI-wheel-compliant
location, even if the package is provided by Anaconda, where
the Qt `PrefixPath` is for example
`C:/Users/<user>/miniconda3/envs/<env-name>/Library`
and the corresponding `LibraryExecutablesPath` is
`C:/Users/<user>/miniconda3/envs/<env-name>/Library/bin`.

Thus, the auto-computed relative target collection directory would
be `bin`, but we need to match the PyPI wheel layout, where the
`LibraryExecutablesPath` is the same as the `PrefixPath`
(i.e., the PySide package directory).