hookutils: qt: fix/extend collection of NSS shared libraries on linux
Fix/extend collection of NSS shared libraries when collecting
`QtWebEngine` on Linux. The existing implementation assumes
that the said shared libraries are located in a `nss` sub-directory
located next to the `libnss3.so` shared library. However, we need
to collect those libraries even if they are not in a separate
directory, but rather located next to the `libnss3.so`.

This is, for example, the case on Debian bookworm:
https://packages.debian.org/bookworm/amd64/libnss3/filelist
whereas previous versions had separate `nss` directory:
https://packages.debian.org/bullseye/amd64/libnss3/filelist