building: extend error message for crashed subprocess during binary dep. analysis
When the isolated subprocess crashes while importing packages prior
to binary dependency analysis (to infer dynamic DLL search paths
on Windows), include the list of planned package imports in the
exception message.

This improves our chances of reproducing the issue when the crash
is related to the order in which packages are imported.