hooks: pandas: exclude conditional import of PyQt5
Add a hook for `pandas.io.clipboard` package to exclude the
conditional import of `PyQt5` that the package makes. This import
is done as a fall-back if `qtpy` is unavailable, however, the
contemporary versions of `pandas` require both `qtpy` and `PyQt5`
for their `clipboard` and `all` extras.

So the `PyQt5` import made in `pandas.io.clipboard` should be
suppressed, to avoid collecting it along with whatver Qt bindings
is selected by our hook for `qtpy`.