analysis: try to pre-load hook as soon as module is encountered
Try to pre-load hooks as soon as their corresponding module is
encountered during the analysis, in an attempt to ensure that
hooks are loaded in roughly the same order as imports are
encountered during the analysis.

This behavior is implicitly required by the new helper
`PyInstaller.utils.hooks.qt.exclude_extraneous_qt_bindings`
for automatic selection of Qt bindings, as it attempts to
infer the preferred Qt bindings based on hooks that have
already been run.

Our hooks are usually loaded in the post-graph analysis phase,
and the order there does not correspond to order of imports
in the code. However, even prior to this commit, some hooks
have been loaded during the analysis stage as well. Specifically,
when a module is imported from another module, our `_safe_import_hook`
implementation checks for exclusion rules in hooks for that
referring (source) module and its parent package(s). This loads
the hooks in question, but because they are cached, they are also
re-used in post-graph processing without being loaded again.

So we augment the hook loading in `_safe_import_hook` by adding
a pre-load step after calling the parent implementation and before
returning the result (i.e., the list of modules that have been
imported).