building: avoid importing pyqtgraph.canvas prior to binary dep. analysis
Add `pyqtgraph.canvas` to the list of packages whose import
should be explicitly suppressed when importing packages for
tracking of dynamic library search path modifications
prior to binary dependency analysis.

The `pyqtgraph.canvas` is known to crash python interpreter
in certain circumstances, and while the issue has been fixed
upstream, we need to work around the issue for older releases
(<= 0.13.3).