building: implement support for setting bytecode optimization level
Implement support for explicitly specifying the bytecode optimization
level for collected python code, independent of the optimization level
in the python process under which PyInstaller is running.

The `Analysis` constructor now accepts an optional argument, `optimize`,
which is propagated through the build process to set the bytecode
optimization level in places where we byte-compile python code
(collected python modules and entry-point script(s)). As this happens
in several low-level places, the existing typecodes in TOC lists have
been extended to encode optimization level (e.g., `PYMODULE-1` and
`PYMODULE-2` in addition to `PYMODULE`).

If `optimize` is not passed to `Analysis` (or is `None` or -1), the
optimization level is taken from `sys.flags.optimize`. However, this is
mostly to accommodate existing spec files. In the newly-generated spec
files, the value is always set based on the command-line options, as
outlined below.

Add optional `--optimize` command-line switch to be used in spec file
generation. If specified, its value is passed to `Analysis` via the
`optimize` argument, and corresponding number of `('O', None, 'OPTION')`
entries is generated for `EXE`, as if they were passed via
`--python-option O` switch(es). If `--optimize` switch is not used,
but `--python-option O` switches are, the optimization level is inferred
from their number. If neither `--optimize` nor `--python-option O` is
used, the optimization level is taken from the python process under
which PyInstaller (or its `pyi-makespec` utility) is running. The
inferred optimization level is written to spec file, thus fixing its
value for subsequent builds that use this spec file (without
regenerating it).