doc: add Bytecode Optimization Level section to manual
Add a section on Bytecode Optimization Level that documents the new
bytecode optimization settings and how to use them.

Add a note in Specifying Python Interpreter Options section to inform
users that using the equivalent of python's -O and -OO option on the
interpreter has no effect on the collected bytecode.