compat: temporarily block import of cffi when importing win32ctypes
`win32ctypes` uses either `ctypes` or `cffi` backend depending on
the availability of `cffi`. The `cffi` backend ends up using
`pycparser`, which requires the docstrings and thus raises error
when python is running in -OO mode (or PYTHONOPTIMIZE is set to 2).

Since availability of `cffi` is determined by an import attempt,
we can "hide" `cffi` by temporarily setting `sys.modules['cffi']`
to `None` and removing the entry (or restore it to its original
value) once we are done importing `win32ctypes` modules.

This work around allows people to run PyInstaller under python with
2nd optimization level, which in turn enables re-use of modulegraph's
code-object cache if the frozen application is also built with
2nd optimization level enabled.