doc: extend documentation on temporary directory for onefile programs
Extend the help message for `--runtime-tmpdir` with a note that
on POSIX, bootloader does not perform environment variable
expansion and that specifying environment variables in the path
will not work.

In `Defining the Extraction Location` section, describe the
possibility of setting environment variable(s) that control
the temporary directory, as an alternative to `--runtime-tmpdir`.
This also documents which environment variables are used under
POSIX and in what order; for Windows, we add reference to
GetTempPathW, whose reference page also documents the temporary
directory selection.