bootloader: remove pyi_path_fullpath_keep_basename
The `pyi_path_fullpath_keep_basename` function was introduced as part
of #3829, which sought to preserve process name of the child process in
onefile applications on POSIX OSes (other than macOS) by disabling the
full symlink resolution when resolving the executable's name.

Subsequently, #5232 re-introduced the executable name resolution via
`readlink()` on `/proc` entry, and implemented process name preservation
via `prctl` call with `PR_GET_NAME` and `PR_SET_NAME`.

At that point, the use of `pyi_path_fullpath_keep_basename˙ became
limited to the fall-back codepath where `/proc` entry resolution failed.

Later on, #7556 explicitly forced the use of this fall-back codepath
when executable is launched via `ld.so` dynamic loader, as resolution
via `/proc` entry ends up pointing to the `ld.so`. In order to handle
cases when the executable is a symbolic link, an explicit resolution
of symbolic link was added at the end of executable path resolution.

At this point, the `pyi_path_fullpath_keep_basename` became completely
redundant; even if it goes out of its way to resolve only symlinks
in the parent path components and leave the file itself alone, if
the file itself is a symbolic link, it gets resolved during the
afore-mentioned final symlink check.

So remove `pyi_path_fullpath_keep_basename` and replace its uses
with calls to `pyi_path_fullpath`, which both canonicalizes the path
and fully resolves the symlinks.

At the same time, fix error messages in `pyi_path_executable` when
`pyi_path_fullpath` fails; several of those seem to be using the
output `execfile` buffer instead of the input buffer that contains
path that is supposed to be resolved.