bootloader: add symlink resolution to common part of pyi_path_executable
Move the explicit symlink resolution step from the end of POSIX
codepath in `pyi_path_executable` to the common codepath at the
end of the function. This effectively adds the explicit symlink
resolution step on Windows, although it is not fully functional
yet due to `pyi_path_fullpath` lacking symlink resolution on
Windows.