archive: splash writer: validate length of strings in 16-byte fields
Ensure that strings that are encoded in fixed-length 16-byte fields
contain 15 or fewer characters, to ensure that there is always a
terminating NULL character present. The current encoding approach
allows the string length to be exactly 16, in which case no zero
padding is performed, and the string is left unterminated, which
would cause issues on the bootloader side.